refactor: modularize web routes
