+++ src/web_app/server.py
@@ -1,15 +1,10 @@
 from __future__ import annotations
 
 import logging
-import shutil
-import uuid
 from pathlib import Path
-import json
-import mimetypes
 
-
-from fastapi import FastAPI, UploadFile, File, HTTPException, Request, Form, Body
-from fastapi.responses import FileResponse, HTMLResponse, PlainTextResponse
+from fastapi import FastAPI, Request
+from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 try:
     from fastapi.templating import Jinja2Templates
@@ -18,12 +13,10 @@ except Exception:  # pragma: no cover
 
 from config import load_config
 from logging_config import setup_logging
-from file_utils import extract_text, merge_images_to_pdf, translate_text
-from file_utils.embeddings import get_embedding, cosine_similarity
-from file_sorter import place_file, get_folder_tree
-import metadata_generation
+from file_utils import extract_text, merge_images_to_pdf, translate_text  # noqa: F401
+import metadata_generation  # noqa: F401
 from . import database
-from models import Metadata, FileRecord, UploadResponse
+from .routes import upload, files, folders, chat
 
 app = FastAPI()
 
@@ -66,488 +59,11 @@ except Exception:
 
 logger = logging.getLogger(__name__)
 
-# --------- Инициализация БД и каталога загрузок ----------
+# --------- Инициализация БД ----------
 database.init_db()
-UPLOAD_DIR = Path("uploads")
-UPLOAD_DIR.mkdir(exist_ok=True)
 
-
-def _resolve_in_output(relative: str) -> Path:
-    """Преобразовать относительный путь в путь внутри output_dir."""
-    base = Path(config.output_dir).resolve()
-    target = (base / relative).resolve()
-    # Защита от выхода за пределы output_dir
-    if not target.is_relative_to(base):
-        raise HTTPException(status_code=400, detail="Path outside output_dir")
-    return target
-
-
-# --------- Маршруты ----------
-@app.post("/upload", response_model=UploadResponse)
-async def upload_file(
-    file: UploadFile = File(...),
-    language: str | None = Form(None),
-    dry_run: bool = False,
-):
-    """Загрузить файл и обработать его."""
-    file_id = str(uuid.uuid4())
-    temp_path = UPLOAD_DIR / f"{file_id}_{file.filename}"
-    try:
-        contents = await file.read()
-        with open(temp_path, "wb") as dest:
-            dest.write(contents)
-
-        # Извлечение текста + генерация метаданных
-        lang = language or config.tesseract_lang
-        text = extract_text(temp_path, language=lang)
-        folder_tree = get_folder_tree(config.output_dir)
-        meta_result = metadata_generation.generate_metadata(text, folder_tree=folder_tree)
-        raw_meta = meta_result["metadata"]
-        if isinstance(raw_meta, dict):
-            metadata = Metadata(**raw_meta)
-        else:
-            metadata = raw_meta
-        metadata.extracted_text = text
-        metadata.language = lang
-        embedding = get_embedding(text)
-
-        meta_dict = metadata.model_dump()
-        # Раскладываем файл по директориям без создания недостающих
-        dest_path, missing = place_file(
-            str(temp_path),
-            meta_dict,
-            config.output_dir,
-            dry_run=dry_run,
-            create_missing=False,
-        )
-        metadata = Metadata(**meta_dict)
-
-    except Exception as exc:  # pragma: no cover
-        logger.exception("Upload/processing failed for %s", file.filename)
-        raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    if missing:
-        database.add_file(
-            file_id,
-            file.filename,
-            metadata,
-            str(temp_path),
-            "pending",
-            meta_result.get("prompt"),
-            meta_result.get("raw_response"),
-            missing,
-            embedding=embedding,
-            suggested_path=str(dest_path),
-        )
-        return UploadResponse(
-            id=file_id,
-            status="pending",
-            missing=missing,
-            suggested_path=str(dest_path),
-        )
-
-    status = "dry_run" if dry_run else "processed"
-
-    # Сохраняем запись в БД
-    database.add_file(
-        file_id,
-        file.filename,
-        metadata,
-        str(dest_path),
-        status,
-        meta_result.get("prompt"),
-        meta_result.get("raw_response"),
-        [],  # missing
-        embedding=embedding,
-        suggested_path=str(dest_path),
-    )
-
-    return UploadResponse(
-        id=file_id,
-        filename=file.filename,
-        metadata=metadata,
-        tags_ru=metadata.tags_ru,
-        tags_en=metadata.tags_en,
-        path=str(dest_path),
-        status=status,
-        missing=[],
-        prompt=meta_result.get("prompt"),
-        raw_response=meta_result.get("raw_response"),
-        suggested_path=str(dest_path),
-    )
-
-
-@app.post("/upload/images", response_model=UploadResponse)
-async def upload_images(
-    files: list[UploadFile] = File(...),
-    language: str | None = Form(None),
-    dry_run: bool = False,
-):
-    """Загрузить несколько изображений, объединить их и обработать как PDF."""
-    file_id = str(uuid.uuid4())
-    temp_dir = UPLOAD_DIR / file_id
-    temp_dir.mkdir(exist_ok=True)
-
-    sorted_files = sorted(files, key=lambda f: f.filename or "")
-    image_paths: list[Path] = []
-    for idx, img in enumerate(sorted_files):
-        temp_img = temp_dir / f"{idx:03d}_{img.filename}"
-        contents = await img.read()
-        with open(temp_img, "wb") as dest:
-            dest.write(contents)
-        image_paths.append(temp_img)
-
-    try:
-        tmp_pdf = merge_images_to_pdf(image_paths)
-        pdf_path = UPLOAD_DIR / f"{file_id}.pdf"
-        shutil.move(tmp_pdf, pdf_path)
-    except Exception as exc:  # pragma: no cover
-        shutil.rmtree(temp_dir, ignore_errors=True)
-        logger.exception("Failed to merge images: %s", [f.filename for f in files])
-        raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    shutil.rmtree(temp_dir, ignore_errors=True)
-
-    try:
-        lang = language or config.tesseract_lang
-        text = extract_text(pdf_path, language=lang)
-        folder_tree = get_folder_tree(config.output_dir)
-        meta_result = metadata_generation.generate_metadata(text, folder_tree=folder_tree)
-        raw_meta = meta_result["metadata"]
-        if isinstance(raw_meta, dict):
-            metadata = Metadata(**raw_meta)
-        else:
-            metadata = raw_meta
-        metadata.extracted_text = text
-        metadata.language = lang
-        embedding = get_embedding(text)
-
-        meta_dict = metadata.model_dump()
-        dest_path, missing = place_file(
-            str(pdf_path),
-            meta_dict,
-            config.output_dir,
-            dry_run=dry_run,
-            create_missing=False,
-        )
-        metadata = Metadata(**meta_dict)
-    except Exception as exc:  # pragma: no cover
-        logger.exception("Upload/processing failed for images")
-        raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    sources = [f.filename for f in sorted_files]
-
-    if missing:
-        database.add_file(
-            file_id,
-            pdf_path.name,
-            metadata,
-            str(pdf_path),
-            "pending",
-            meta_result.get("prompt"),
-            meta_result.get("raw_response"),
-            missing,
-            sources=sources,
-            embedding=embedding,
-            suggested_path=str(dest_path),
-        )
-        return UploadResponse(
-            id=file_id,
-            status="pending",
-            missing=missing,
-            sources=sources,
-            suggested_path=str(dest_path),
-        )
-
-    status = "dry_run" if dry_run else "processed"
-    database.add_file(
-        file_id,
-        pdf_path.name,
-        metadata,
-        str(dest_path),
-        status,
-        meta_result.get("prompt"),
-        meta_result.get("raw_response"),
-        [],
-        sources=sources,
-        embedding=embedding,
-        suggested_path=str(dest_path),
-    )
-
-    return UploadResponse(
-        id=file_id,
-        filename=pdf_path.name,
-        metadata=metadata,
-        tags_ru=metadata.tags_ru,
-        tags_en=metadata.tags_en,
-        path=str(dest_path),
-        status=status,
-        missing=[],
-        prompt=meta_result.get("prompt"),
-        raw_response=meta_result.get("raw_response"),
-        sources=sources,
-        suggested_path=str(dest_path),
-    )
-
-
-@app.get("/metadata/{file_id}", response_model=Metadata)
-async def get_metadata(file_id: str):
-    record = database.get_file(file_id)
-    if not record:
-        raise HTTPException(status_code=404, detail="Metadata not found")
-    return record.metadata
-
-
-@app.get("/download/{file_id}")
-async def download_file(file_id: str, lang: str | None = None):
-    record = database.get_file(file_id)
-    if not record:
-        raise HTTPException(status_code=404, detail="File not found")
-
-    if lang:
-        extracted = record.metadata.extracted_text or ""
-        orig_lang = record.metadata.language
-        if lang == orig_lang:
-            text = extracted
-        elif record.translation_lang == lang and record.translated_text:
-            text = record.translated_text
-        else:
-            text = translate_text(extracted, lang)
-            database.update_file(
-                file_id,
-                translated_text=text,
-                translation_lang=lang,
-            )
-        headers = {
-            "Content-Disposition": f"attachment; filename={record.filename}_{lang}.txt"
-        }
-        return PlainTextResponse(text, headers=headers)
-
-    path = Path(record.path)
-    if not path.exists():
-        raise HTTPException(status_code=404, detail="File not found")
-    return FileResponse(path, filename=path.name)
-
-
-@app.get("/preview/{file_id}")
-async def preview_file(file_id: str):
-    record = database.get_file(file_id)
-    if not record:
-        raise HTTPException(status_code=404, detail="File not found")
-    path = Path(record.path)
-    if not path.exists():
-        raise HTTPException(status_code=404, detail="File not found")
-    content_type, _ = mimetypes.guess_type(path.name)
-    return FileResponse(path, media_type=content_type or "application/octet-stream")
-
-
-@app.get("/files/{file_id}/details", response_model=FileRecord)
-async def get_file_details(file_id: str, lang: str | None = None):
-    record = database.get_details(file_id)
-    if not record:
-        raise HTTPException(status_code=404, detail="File not found")
-    if lang:
-        extracted = record.metadata.extracted_text or ""
-        orig_lang = record.metadata.language
-        if lang == orig_lang:
-            text = extracted
-        elif record.translation_lang == lang and record.translated_text:
-            text = record.translated_text
-        else:
-            text = translate_text(extracted, lang)
-            database.update_file(
-                file_id,
-                translated_text=text,
-                translation_lang=lang,
-            )
-            record.translated_text = text
-            record.translation_lang = lang
-        record.translated_text = record.translated_text or text
-        record.translation_lang = lang
-    return record
-
-
-@app.get("/files", response_model=list[FileRecord])
-async def list_files():
-    return database.list_files()
-
-
-@app.get("/search/semantic")
-async def semantic_search(q: str):
-    """Семантический поиск по сохранённым документам."""
-    query_vec = get_embedding(q)
-    results: list[dict[str, object]] = []
-    for rec in database.list_files():
-        emb = rec.embedding
-        if not emb:
-            continue
-        score = cosine_similarity(query_vec, emb)
-        results.append({"id": rec.id, "filename": rec.filename, "score": score})
-    results.sort(key=lambda x: x["score"], reverse=True)
-    return {"results": results}
-
-
-@app.patch("/files/{file_id}", response_model=FileRecord)
-async def update_file(file_id: str, data: dict = Body(...)):
-    record = database.get_file(file_id)
-    if not record:
-        raise HTTPException(status_code=404, detail="File not found")
-    metadata_updates = data.get("metadata") or {}
-    new_metadata_dict = record.metadata.model_dump()
-    if metadata_updates:
-        new_metadata_dict.update(metadata_updates)
-    new_metadata = Metadata(**new_metadata_dict)
-    path_param = data.get("path")
-    status = data.get("status")
-    prompt = data.get("prompt")
-    raw_response = data.get("raw_response")
-    missing = data.get("missing")
-
-    old_path = Path(record.path)
-    if not old_path.exists():
-        raise HTTPException(status_code=404, detail="File not found on disk")
-    dest_path = old_path
-
-    if metadata_updates and not path_param:
-        old_json = old_path.with_suffix(old_path.suffix + ".json")
-        if old_json.exists():
-            old_json.unlink()
-        dest_path, _ = place_file(
-            old_path,
-            new_metadata_dict,
-            config.output_dir,
-            dry_run=False,
-            create_missing=True,
-        )
-        new_metadata = Metadata(**new_metadata_dict)
-    elif path_param:
-        old_json = old_path.with_suffix(old_path.suffix + ".json")
-        if old_json.exists():
-            old_json.unlink()
-        dest_path = _resolve_in_output(path_param)
-        dest_path.parent.mkdir(parents=True, exist_ok=True)
-        if dest_path.exists():
-            raise HTTPException(status_code=409, detail="Target already exists")
-        shutil.move(str(old_path), dest_path)
-        with open(dest_path.with_suffix(dest_path.suffix + ".json"), "w", encoding="utf-8") as f:
-            json.dump(new_metadata_dict, f, ensure_ascii=False, indent=2)
-
-    database.update_file(
-        file_id,
-        metadata=new_metadata if metadata_updates else None,
-        path=str(dest_path) if (metadata_updates or path_param) else None,
-        status=status,
-        prompt=prompt,
-        raw_response=raw_response,
-        missing=missing,
-    )
-    return database.get_file(file_id)
-
-
-@app.get("/folder-tree")
-async def folder_tree():
-    """Вернуть структуру папок в выходном каталоге."""
-    return get_folder_tree(config.output_dir)
-
-
-@app.post("/folders")
-async def create_folder(path: str):
-    """Создать каталог внутри output_dir."""
-    target = _resolve_in_output(path)
-    if target.exists():
-        raise HTTPException(status_code=409, detail="Folder already exists")
-    target.mkdir(parents=True, exist_ok=False)
-    return get_folder_tree(config.output_dir)
-
-
-@app.patch("/folders/{folder_path:path}")
-async def rename_folder(folder_path: str, new_name: str):
-    src = _resolve_in_output(folder_path)
-    if not src.exists():
-        raise HTTPException(status_code=404, detail="Folder not found")
-    dest_relative = str(Path(folder_path).parent / new_name)
-    dest = _resolve_in_output(dest_relative)
-    if dest.exists():
-        raise HTTPException(status_code=409, detail="Target already exists")
-    src.rename(dest)
-    return get_folder_tree(config.output_dir)
-
-
-@app.delete("/folders/{folder_path:path}")
-async def delete_folder(folder_path: str):
-    target = _resolve_in_output(folder_path)
-    base = Path(config.output_dir).resolve()
-    if target == base:
-        raise HTTPException(status_code=400, detail="Cannot delete root folder")
-    if not target.exists():
-        raise HTTPException(status_code=404, detail="Folder not found")
-    shutil.rmtree(target)
-    return get_folder_tree(config.output_dir)
-
-
-# ---------- ЕДИНЫЙ finalize ----------
-@app.post("/files/{file_id}/finalize", response_model=FileRecord)
-async def finalize_file(
-    file_id: str,
-    missing: list[str] | None = Body(default=None),
-):
-    """
-    Завершить обработку «отложенного» файла:
-    - при наличии `missing` — создать указанные каталоги внутри output_dir;
-    - перенести файл в целевую структуру (create_missing=True);
-    - обновить запись в БД.
-    """
-    record = database.get_file(file_id)
-    if not record:
-        raise HTTPException(status_code=404, detail="File not found")
-    if record.status != "pending":
-        # Возвращаем текущую запись, ничего не делаем
-        return record
-
-    # Если фронт подсказал недостающие, создадим их тут (без дергания собственного эндпоинта)
-    if missing:
-        for rel in missing:
-            target = _resolve_in_output(rel)
-            target.mkdir(parents=True, exist_ok=True)
-
-    # Путь к временно загруженному файлу
-    temp_path = record.path
-    if not temp_path:
-        temp_path = str(UPLOAD_DIR / f"{file_id}_{record.filename}")
-
-    meta_dict = record.metadata.model_dump()
-    dest_path, still_missing = place_file(
-        str(temp_path),
-        meta_dict,
-        config.output_dir,
-        dry_run=False,
-        create_missing=True,
-    )
-    metadata = Metadata(**meta_dict)
-
-    database.update_file(
-        file_id,
-        metadata,
-        str(dest_path),
-        "processed",
-        record.prompt,
-        record.raw_response,
-        still_missing,
-        suggested_path=str(dest_path),
-    )
-    return database.get_file(file_id)
-
-
-@app.post("/chat/{file_id}")
-async def chat(file_id: str, message: str = Body(..., embed=True)):
-    """Простой чат с учётом текста файла."""
-    record = database.get_file(file_id)
-    if not record:
-        raise HTTPException(status_code=404, detail="File not found")
-
-    text = record.metadata.extracted_text or ""
-    reply = f"Эхо: {message} | {text}".strip()
-
-    database.add_chat_message(file_id, "user", message)
-    history = database.add_chat_message(file_id, "assistant", reply)
-    return {"response": reply, "chat_history": history}
+# --------- Подключение маршрутов ----------
+app.include_router(upload.router)
+app.include_router(files.router)
+app.include_router(folders.router)
+app.include_router(chat.router)
